backend_flask/requirements.txt
Flask
Flask-Bcrypt
Flask-Login
python-dotenv
Pillow
transformers
torch
scikit-learn
openai
google-generativeai
requests
spacy
numpy
pandas
tqdm
